convenient changes

+++ main.py
@@ -15,23 +15,34 @@ data={}
 
 def Load():
     global data
-    with open(path_json, "r", encoding="utf-8_sig") as f:
-        data = json.load(f)
-        Initialize()
-        print("json file loaded")
+    if not os.path.exists(path_json):
+        with open(path_json, "w", encoding="utf-8_sig") as f:
+            json.dump({}, f, indent=4, ensure_ascii=False)
+            LogSys(0,"json file created")
+    try:
+        with open(path_json, "r", encoding="utf-8_sig") as f:
+            data = json.load(f)
+            Initialize()
+            LogSys(0,"json loaded")
+    except Exception as e:
+        LogSys(2,"json load failed")
+        print(f"{type(e)} : {str(e)}")
 
 def Save():
     global data 
-    with open(path_json, "w", encoding="utf-8_sig") as f:
-        json.dump(data, f, indent=4, ensure_ascii=False)
-        print("json file saved")
+    try:
+        with open(path_json, "w", encoding="utf-8_sig") as f:
+            json.dump(data, f, indent=4, ensure_ascii=False)
+            LogSys(0,"json saved")
+    except Exception as e:
+        LogSys(2,"json file save failed")
+        print(f"{type(e)} : {str(e)}")
 
 def Initialize():
     global data
     dists=["notice_group"]
-    vars=["target_forum"]
+    vars=["target_forum", "log_channel"]
     for dist in dists:
-
         if dist not in data:
             data[dist] = {}
     for var in vars:
@@ -40,6 +51,7 @@ def Initialize():
     Save()
 
 Load()
+time.timezone = 32400 # JST
 
 #region 絵文字の判定
 def is_discord_emoji(s: str) -> bool:
@@ -53,11 +65,25 @@ def is_unicode_emoji(s: str) -> bool:
 
 #region 便利関数
 
-async def Reply(interaction: discord.Integration, type:int, title: str, message: str, public: bool = True):
+async def Reply(itr: discord.Integration, type:int, title: str, message: str, public: bool = True):
     """type: {0:成功,1:情報,2:エラー}"""
     colors = [discord.Color.green(), discord.Color.blue(), discord.Color.red()]
     emb = discord.Embed(title=title, description=message, color=colors[type])
     await Reply(itr, 2, "エラー", embed=emb, ephemeral=not public)
+
+async def LogCh(channel_id, string: str):
+    """指定されたスレッドにメッセージを送信します"""
+    channel = await bot.get_channel(int(channel_id))
+    if channel is not None:
+        try:
+            await channel.send(string)
+        except discord.Forbidden:
+            pass
+
+def LogSys(type:int, string: str):
+    """type: {0:成功, 1:情報, 2:エラー, 3:その他}"""
+    colors = ["\033[32", "\033[36", "\033[31", "\033[37"]
+    print(f"[{time.strftime('%Y/%m/%d %H:%M:%S')}] {str(type)} | {colors[type]}{string}\033[0m")
 
 #region イベント
 @bot.event
@@ -92,7 +118,7 @@ async def on_reaction_add(reaction, user):
 #region コマンド
 
 @tree.command(name='add_thread', description="絵文字に対応するスレッドを作成します")
-@tree.descrive(emoji = "絵文字1文字", thread_name = "スレッド名")
+@app_commands.describe(emoji = "絵文字1文字", thread_name = "スレッド名")
 async def add_thread(itr: discord.Interaction, emoji: str, thread_name: str):
     global data
     if data["target_forum"] == "":
@@ -111,15 +137,16 @@ async def add_thread(itr: discord.Interaction, emoji: str, thread_name: str):
                 if forum is None:
                     await Reply(itr, 2, "エラー", "フォーラムが見つかりませんでした")
                 else:  
-                    data["notice_group"][emoji] = {}
-                    _temp ={"owner": str(itr.user.id),
-                            "thread_id": str(thread.id),
-                            "created_at": str(time.time()),
-                            "messages":{}
-                           }
+                    data["notice_group"][emoji] = {
+                        "owner": str(itr.user.id),
+                        "thread_id": str(thread.id),
+                        "created_at": str(time.time()),
+                        "messages":{}
+                    }
                     await Reply(itr, 2, "エラー", "")
+
 @tree.command(name='remove_thread', description="絵文字に対応するスレッドを削除します")
-@tree.describe(emoji = "絵文字1文字")
+@app_commands.describe(emoji = "絵文字1文字")
 async def remove_thread(itr: discord.Interaction, emoji: str):
     global data
     if not emoji in data["notice_group"]:
@@ -141,38 +168,48 @@ async def remove_thread(itr: discord.Interaction, emoji: str):
         except:
             await Reply(itr,2, "エラー", "スレッドの削除に失敗しました。")
                 
-# @tree.command(name='expire', description="スレッド内のメッセージの有効期限を設定できます")
-# @tree.descrive(msg_link = "**転送された**メッセージのリンク", expire_at = "有効期限 (YYYY/MM/DD HH:MM の書式)")
-# async def expire(itr: discord.Interaction, msg_link: str, expire_at: str):
-#     try:
-#         msg = await commands.MessageConverter(msg_link).convert(itr)
-#     except commands.MessageNotFound:
-#         itr.command_failed = True
-#         await Reply(itr,2, "エラー", "メッセージの取得に失敗しました")
-#     except:
-#         itr.command_failed = True
-#         await Reply(itr,2, "エラー", "例外が発生しました")
-    
-#     if re.fullmatch(r"\d{4}/\d{2}/\d{2} \d{2}:\d{2}", expire_at):
-#         expire_at = time.mktime(time.strptime(expire_at, "%Y/%m/%d %H:%M"))
-#         if expire_at < time.time():
-#             itr.command_failed = True
-#             await Reply(itr,2, "エラー", "有効期限が過去の時間です")
-#         else:
-#             for emoji in data["notice_group"]:
-#                 for message in data["notice_group"][emoji]["messages"]:
-#                     if message == str(msg.id):
-#                         data["notice_group"][emoji]["messages"][message]["expire_at"] = str(expire_at)
-#                         break
-#             Save()
-#             await Reply(itr,0, "成功", f"メッセージの有効期限を{expire_at}に設定しました")
-#     elif re.fullmatch(r"\d{4}/\d{2}/\d{2}", expire_at):
-    
-    
-    
-    
+@tree.command(name='expire', description="スレッド内のメッセージの有効期限を設定できます")
+@app_commands.describe(msg_link = "**転送された**メッセージのリンク", expire_at = "有効期限 (YYYY/MM/DD HH:MM or YYYY/MM/DD の書式)")
+async def expire(itr: discord.Interaction, msg_link: str, expire_at: str):
+    try:
+        msg = await commands.MessageConverter(msg_link).convert(itr)
+        if re.fullmatch(r"\d{4}/\d{2}/\d{2} \d{2}:\d{2}", expire_at) or re.fullmatch(r"\d{4}/\d{2}/\d{2}", expire_at):
+            if re.fullmatch(r"\d{4}/\d{2}/\d{2} \d{2}:\d{2}", expire_at):
+                expire_at = time.mktime(time.strptime(expire_at, "%Y/%m/%d %H:%M"))
+            else:
+                expire_at = time.mktime(time.strptime(expire_at, "%Y/%m/%d"))
+                expire_at += 86400 # 1日後に設定(翌日になったら削除)
+            
+            if expire_at < time.time():
+                itr.command_failed = True
+                await Reply(itr,2, "エラー", "有効期限が過去の時間です")
+            else:
+                is_found = False
+                for emoji in data["notice_group"]:
+                    for message in data["notice_group"][emoji]["messages"]:
+                        if data["notice_group"][emoji]["messages"][message]["forwarded_msg_id"] == str(msg.id):
+                            data["notice_group"][emoji]["messages"][message]["expire_at"] = str(expire_at)
+                            is_found = True
+                            Save()
+                            break
+                if is_found:
+                    await Reply(itr,0, "成功", f"メッセージの有効期限を{expire_at}に設定しました")
+                else:
+                    itr.command_failed = True
+                    await Reply(itr,2, "エラー", "そのメッセージは転送されたものではありません\nスレッドに転送されたメッセージのリンクを指定してください")
+        else:
+            itr.command_failed = True
+            await Reply(itr,2, "エラー", "有効期限の書式が間違っています\nYYYY/MM/DD HH:MM または YYYY/MM/DD の書式で指定してください")
+            
+    except commands.MessageNotFound:
+        itr.command_failed = True
+        await Reply(itr,2, "エラー", "メッセージの取得に失敗しました")
+    except:
+        itr.command_failed = True
+        await Reply(itr,2, "エラー", "例外が発生しました")
+
 @tree.command(name='stats', description="指定されたボイスチャットチャンネルの状態を確認できます")
-@tree.describe(channel = "ボイスチャンネル")
+@app_commands.describe(channel = "ボイスチャンネル")
 async def stats(itr: discord.Interaction, channel: discord.VoiceChannel):
     members_with_bot = 0
     members_without_bot = 0
@@ -194,8 +231,8 @@ async def stats(itr: discord.Interaction, channel: discord.VoiceChannel):
             members_all_muted += 1
     await Reply(itr, 1, f"{channel.mention} の状態", f"通話中の人数:{members_with_bot}\n通話中の人数(Botを除く):{members_without_bot}\nBotの数:{members_bot}\nミュート中の人数:{members_muted}\nスピーカーミュート中の人数:{members_deafen}\n全ミュートの人数:{members_all_muted}")
     
-@tree.command(name='set_forum',description="このボットがメインで動くフォーラムを指定します")
-@tree.describe(channel = "フォーラムチャンネル")
+@tree.command(name='set_forum', description="このボットがメインで動くフォーラムを指定します")
+@app_commands.describe(forum = "フォーラムチャンネル")
 async def set_forum(itr: discord.Interaction, forum: discord.ForumChannel):
     if not itr.user.id in admins:
         itr.command_failed = True
@@ -209,7 +246,19 @@ async def set_forum(itr: discord.Interaction, forum: discord.ForumChannel):
 
 @tasks.loop(seconds=15)
 async def Check_expires():
-    print()
+    global data
+    now = time.time()
+    for emoji in data["notice_group"]:
+        for message in data["notice_group"][emoji]["messages"]:
+            if "expire_at" in data["notice_group"][emoji]["messages"][message]:
+                if now > float(data["notice_group"][emoji]["messages"][message]["expire_at"]):
+                    try:
+                        msg : discord.Message = await commands.MessageConverter(data["notice_group"][emoji]["messages"][message]["forwarded_msg_id"]).convert(bot)
+                        await msg.delete()
+                        del data["notice_group"][emoji]["messages"][message]
+                        Save()
+                    except:
+                        pass
 
 # token = os.getenv("DISCORD_TOKEN")
 token = input("Tokenを入力>>>")
